- updated permissions for commands - updated help command itself
