+++ src/Commands/Server Information/UserInfoCommand.ts
@@ -23,7 +23,7 @@ export class UserInfoCommand extends Command {
             new CommandPermission(
                 [],
                 [],
-                [],
+                ["suspended"],
                 [],
                 true
             ),
